separate App from index
